# book-review-api/src/main/resources/static/openapi/openapi.yml
openapi: 3.0.3
info:
  title: 書籍レビューAPI
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /books:
    get:
      tags:
      - book-controller
      operationId: getBooks
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookResponse"
    post:
      tags:
        - book-controller
      operationId: createBook

      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - author
                - images
              properties:
                title:
                  type: string
                author:
                  type: string
                images:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"
  /books/{bookId}/reviews:
    get:
      tags:
      - review-controller
      operationId: getReviews
      parameters:
      - name: bookId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewResponse"
    post:
      tags:
      - review-controller
      operationId: createReview
      parameters:
      - name: bookId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewResponse"
  /books/{bookId}/images:
    post:
      tags:
      - book-image-controller
      operationId: upload
      parameters:
      - name: bookId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
              required:
              - image
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookImageResponse"
  /books/{bookId}:
    get:
      tags:
      - book-controller
      operationId: getBook
      parameters:
      - name: bookId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"
    delete:
      tags:
      - book-controller
      operationId: deleteBook
      parameters:
      - name: bookId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
    patch:
      tags:
      - book-controller
      operationId: updateBook
      parameters:
      - name: bookId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"
  /books/{bookId}/reviews/{reviewId}:
    get:
      tags:
      - review-controller
      operationId: getReview
      parameters:
      - name: bookId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewResponse"
    delete:
      tags:
      - review-controller
      operationId: deleteReview
      parameters:
      - name: bookId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
    patch:
      tags:
      - review-controller
      operationId: updateReview
      parameters:
      - name: bookId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewResponse"
  /books/{bookId}/images/{imageId}:
    delete:
      tags:
      - book-image-controller
      operationId: delete
      parameters:
      - name: bookId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: imageId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
components:
  schemas:
    BookImageResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        imageUrl:
          type: string
    BookResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        author:
          type: string
        images:
          type: array
          items:
            $ref: "#/components/schemas/BookImageResponse"
    ReviewRequest:
      type: object
      properties:
        content:
          type: string
        rating:
          type: integer
          format: int32
    ReviewResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        bookId:
          type: integer
          format: int64
        content:
          type: string
        rating:
          type: integer
          format: int32
    BookRequest:
      type: object
      properties:
        title:
          type: string
        author:
          type: string